+++ Sources/NetworkClient/Protocols/InterceptorProtocol.swift
@@ -8,18 +8,21 @@
 import Foundation
 
 
-public protocol InterceptorProtocol: AnyObject {
+public protocol InterceptorProtocol: Sendable {
     typealias Result = NetworkClient.ChainResult
     
     func process(
         _ request: URLRequest,
-        completion: (URLRequest) async throws -> NetworkClient.ChainResult
+        next: @Sendable (URLRequest) async throws -> NetworkClient.ChainResult
     ) async throws -> NetworkClient.ChainResult
 }
 
-public class NoOpAdapter : InterceptorProtocol {
-    public func process(_ request: URLRequest, completion: (URLRequest) async throws -> NetworkClient.ChainResult) async throws -> NetworkClient.ChainResult {
-        try await completion(request)
+public struct NoOpAdapter : InterceptorProtocol {
+    public func process(
+        _ request: URLRequest,
+        next: @Sendable (URLRequest) async throws -> NetworkClient.ChainResult
+    ) async throws -> NetworkClient.ChainResult {
+        try await next(request)
     }
 }