+++ Sources/NetworkUtilities/MimeType.swift
@@ -7,7 +7,10 @@
 
 
 import UniformTypeIdentifiers
+
+#if os(iOS)
 import MobileCoreServices
+#endif
 
 internal extension URL {
     var mimeType: String? {
@@ -15,7 +18,7 @@ internal extension URL {
             return nil
         }
         
-        if #available(iOS 14, *),
+        if #available(iOS 14, macOS 11.0, *),
            let mime = UTType(filenameExtension: pathExtension)?.preferredMIMEType
         {
             return mime